# backend/app/routers/courses.py
import asyncio
from collections.abc import Iterator
from typing import Any

import httpx
import logfire
from fastapi import APIRouter
from pydantic import BaseModel, ConfigDict, Field, TypeAdapter

from ..dependencies import CanvasAuth, HTTPClient
from ..parsers.base import PlannerNote

router = APIRouter(prefix="/courses")

class Course(BaseModel):
    model_config = ConfigDict(extra="allow")

    id: int
    name: str
    course_code: str

 # MAYBE: sort by date modified in future?(might add too many api calls)
class CourseFile(BaseModel):
    model_config = ConfigDict(extra="allow")

    file_id: int = Field(validation_alias="content_id")
    item_id: int = Field(validation_alias="id") # might remove later if still not needed
    title: str

# MAYBE: include IDs later
class BulkDeleteResult(BaseModel):
    total: int
    deleted: int
    failed: int


@router.get("")
async def get_courses(client: HTTPClient, auth: CanvasAuth, inactive: bool = False) -> list[Course]:
    params: dict[str, Any] = {"per_page": 100}
    if not inactive:
        params["enrollment_state"] = "active"
    courses_resp = await client.get("/courses", headers=auth, params=params)
    courses_resp.raise_for_status()
    return courses_resp.json()

# Maybe add query parameter to limit to schedules
@router.get("/{course_id}/pdfs", summary="List course PDFs")
async def get_pdfs(client: HTTPClient, auth: CanvasAuth, course_id: int) -> list[CourseFile]: # TODO: Strengthen include parameter
    params: dict[str, Any] = {"include": ["items"], "per_page": 100}
    resp = await client.get(f"/courses/{course_id}/modules", headers=auth, params=params)
    resp.raise_for_status()
    return [
        file
        for _, file in sorted(iter_files(resp.json()), key=lambda t: t[0])
        if file.title.lower().endswith((".pdf", ".docx"))
    ]

def iter_files(modules: list[dict[str, Any]]) -> Iterator[tuple[tuple[int, int], CourseFile]]:
    for module in modules:
        for item in module["items"]:
            if item["type"] == "File":
                yield (module["position"], item["position"]), CourseFile.model_validate(item)


_planner_notes_adapter = TypeAdapter(list[PlannerNote])

async def delete_note(client: httpx.AsyncClient, auth: CanvasAuth, note: PlannerNote) -> bool:
    try:
        resp = await client.delete(f"/planner_notes/{note.id}", headers=auth)
        resp.raise_for_status()
        return True
    except httpx.HTTPError:
        logfire.warning(
            "Failed to delete note: {title}",
            title=note.title, date=note.todo_date, note_id=note.id,
            course_id=note.course_id, _exc_info=True,
        )
        return False

# TODO: Add semaphore
@router.delete("/{course_id}/notes", summary="Delete all planner notes for a course") # might not be idiomatic for webapp
async def delete_notes(client: HTTPClient, auth: CanvasAuth, course_id: int) -> BulkDeleteResult:
    params = {"context_codes[]": f"course_{course_id}", "per_page": 100}
    resp = await client.get("/planner_notes", params=params, headers=auth)
    resp.raise_for_status()
    # resp.content to use rust validate_json speedup(canvas api is utf-8)
    planner_notes: list[PlannerNote] = _planner_notes_adapter.validate_json(resp.content)
    results = await asyncio.gather(*(delete_note(client, auth, note) for note in planner_notes))

    total = len(planner_notes)
    deleted = sum(results)
    failed = total - deleted
    logfire.info(
        "Deleted planner notes for course {course_id}",
        course_id=course_id,
        total=total,
        deleted=deleted,
        failed=failed,
    )
    return BulkDeleteResult(total=total, deleted=deleted, failed=failed)